Centralizing the Document Helpers. Also fixing issues with the new extension for MongoDBBson()
